+++ src/main/kotlin/com/github/janneri/innerbuildergeneratorintellijplugin/GeneratorOptions.kt
@@ -3,5 +3,6 @@ package com.github.janneri.innerbuildergeneratorintellijplugin
 data class GeneratorOptions(
     val generateCopyMethod: Boolean,
     val methodPrefix: String,
-    val jsonDeserializeWithBuilder: Boolean
+    val paramName: String,
+    val jsonDeserializeWithBuilder: Boolean,
 )